backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
openai==1.30.1
structlog==24.1.0
prometheus-client==0.20.0
psycopg2-binary
